OM-36138: Add support for using "key" fields in the stat request "groupBy" clause.
This is intended to fix an issue where the presentation of the CPU Allocation and Memory Allocation stats are presented incorrectly in the cluster overview UI. This bug was because the groupBy "key" wasn't having any effect in the stats retrieval, and since the allocation stats seem to be the only PM stats that have commodity keys set on them, this resulted in the cluster memory allocation stats showing the stats for a single entity in the cluster.

This only adds support for grouping by commodity key. Grouping by related entities or by virtual disk will be handled in a separate issue, OM 36453.