OM-35979: XL Repo: Pagination for projected plan per-entity stats.
This review adds pagination to the RepositoryService interface to get per-entity plan stats.

The pagination is done after the topology is read from Arango, and the stats are extracted from the TopologyEntityDTO's. We already have an EntityStatsPaginator that can apply in-memory pagination to a list of EntityStats objects, so we are re-using it in the repository.

In the future, if we keep Arango, it may be worth it to store commodities for plans separately to speed up lookups.

Not done: Pagination by price index. We don't currently store the per-entity price index. This will be handled by: om36437

Testing Done:
- Unit tests, with additional test.

Reviewed By: zhengbin
https://rbcommons.com/s/VMTurbo/r/24374