OM-36659: XL: StorageStress Probe will not start RB-24482: Patrick
Background:
The storagestressprobe is an SDK probe that is part of the sdk.examples package for simulating a storage target.
The "mediation-storagestressprobe" component in XL runs the storagestressprobe in a container.

Problem:
The mediation-storagestressprobe does not start in the current XL environment. It results in an error
```
System or environment property "server_port" must be set
```

Solution:
add the definitions required for spring configuration that were added during the removal of spring-boot.