OM-36511: XL: Set memberCount and activeEntityCount when mapping groups to API DTOs.
Issue: groups always appear to have 0 entities, and when scoping to a "group" the "List of <>" tab doesn't appear.

We currently just set "entityCount", but the API definitions mainly use "memberCount" to determine the number of members in a group. Entity count is now reserved to groups-of-groups, to determine the total number of entities in a parent group.

I'm also setting activeEntityCount, but not querying the actual entities to get the right numbers. We can do that later, more efficiently (opened bug 36862).

Note: There is an issue where the "Actions" tab also intermittently doesn't appear. I think this is some kind of UI state issue, but we will investigate it more.

Testing Done:
- Started up XL.
- Verified the List of Virtual Machines tab appears in a group of VMs, both static and dynamic.

Reviewed By: karthikt
https://rbcommons.com/s/VMTurbo/r/24604/