OM-36913: XL: Add support for /doc route to enable local development/debugging of UI on top of XL.
The UI recently moved documentation and all doc-related scripts (and the vmturbo-doc module) to a /doc path instead of /app/doc. When un-bundled for a local development build (e.g. using gulp watch) this breaks running the UI on top of XL, because all requests to /doc return a NOT FOUND message. This change adds support for /doc to the external API servlet and the ngingx container.

Testing Done:
- I was unable to run a local debuggable UI before the change, and I can run it afterwards.

Reviewed By: karthikt, psoohoo
https://rbcommons.com/s/VMTurbo/r/24627/