OM-36455: XL: Set Storage entities' Storage Access Capacity from IOPS Setting if provider capacity not available.
In XL, Storage entities would sometimes come out of stitching with Storage Access sold capacity 0 (which is illegal).

The reason this happened is because in Hypervisor probes Storage entities do not buy Storage Access from the underlying disk arrays/logical pools, so there was no provider capacity available to set the storage acccess capacity on the Storage.

The fix (mirroring Classic) is to add an IOPSCapacity setting for "Storage" entities, and use that setting value when there is no provider capacity.

The final priority order for Storage Access capacity in storage entities is:
1) The capacity of the underlying disk array/logical pool, if available.
2) The capacity set on the Storage entity by the probe, if available.
3) The capacity derived from the IOPSCapacity setting for the storage entity.
4) The hacked large value!

Testing Done:
- Unit tests, including additional unit tests.
- Brought up XL with vsphere-dc13 and verified storage capacity is set to the value of the IOPSCapacity setting.

Reviewed By: dblinn
https://rbcommons.com/s/VMTurbo/r/24692/