OM-37177: XL: Send price index alongside the projected topology instead of separately.
Right now we send the price index and projected topology separately, which makes it significantly more complicated to process the messages (i.e. have to worry about interleaving, getting one but not the other, inconsistencies between the messages, and so on).

This change sends the price index together with the projected topology. TopologyEntityDTOs are now wrapped in a ProjectedTopologyEntity message, which contains the entity and it's original and projected price index. Making all the relevant changes in the history component and repository to process the new data format.

Note: The handling of skipped entities is now done in the history component. The Market sends the list of skipped entities, and the history is responsible for setting their original price index to the default (i.e. 1).

Testing Done:
- Full build.
- System testing:
   - Brought up XL.
   - Tested running a plan (viewing optimized improvements widget), looking at the stats in the UI, and checking that price index rows exist in the database.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/24592/