OM-37109: XL Market: Add a MarketDebugService and AnalysisDebuggingTest to help ease market debugging workflow.
Intent:
To enable easier end-to-end debugging of the Market component on development/test instances of XL.

Background:
It's currently difficult to debug market-related issues in XL, especially if we want to step through market code in the debugger. We need to connect to the appliance, and trigger a broadcast each time. Furthermore, the appliance doesn't have sources, making the debugging even more difficult.

This change adds a MarketDebugService (enabled in development/test environments - most notably the local dev environment) that can be used to retrieve the inputs and outputs of the most recent completed analyses (by context ID) from the market component. It also includes an AnalysisDebuggingTest which sets up the harness for:
1) Reading the input from a JSON file.
2) Running the Analysis
3) Looking at the results on a per-entity basis.

See the javadoc in AnalysisDebuggingTest for more info on how it should work.

To help support running the analysis from the input, I'm also moving the retrieval of global settings out of the Analysis and into the MarketRunner.

Testing Done:
- Full build.
- Used MarketDebugService to debug market issue.

Reviewed By: dblinn
https://rbcommons.com/s/VMTurbo/r/24750/