OM-36736: XL History/API: Add global entity type optimization to historical per-entity stats RPC.
Intent:
To speed up queries like "get per-entity stats for all entities of type X in the realtime topology." These are used, for example, by the optimized improvements widget.

Background:
Right now, when handling queries for aggregated stats for all entities in the market (or all entities in a global-scope temporary group in the market) there are several inefficiencies:
1) We use a remote call to the repository to get the entity IDs that the scopes resolve to.
2) We send the list of entity IDs to the history component (this list can be thousands of entities long)
3) We send the list of entity IDs to the database.

The global entity type optimization omits all three of those steps - we just send the entity type to the history component, and the history component reads the top X entities from that entity type's table.

Changes:
Using a new "EntityStatScope" message, which is either a list of explicit OIDs or an entity type, in the GetEntityStats RPC. In the history component this propagates down to HistoryDbIO, and in the API component we construct the message by checking for the two special conditions - a single scope of "Market" with an entity type, or a single scope of a global temp group.

Testing Done:
- Full build, with new unit test.
- System testing - brought up XL, verified optimized improvement widget functionality.
    - Didn't test the difference in time in a large environment.

Reviewed By: zhengbin
https://rbcommons.com/s/VMTurbo/r/24843/