OM-36622: XL: Use GetEntityStats API to sort entity IDs by utilization, instead of a separate RPC call.
Intent: Reduce redundancy and clean up the StatsService API.

Background:
The GetPaginationEntityByUtilization RPC call was added as part of the search pagination work to sort a list of IDs (or the global entity type) by price index. Back then history pagination was not fully implemented, so it made sense to use this specialized API for pagination all the way down to the database.

Now that we have historical pagination all the way down to the database, we can use a request to GetEntityStats to get the same information (EntityStats objects have an OID, and come back ordered by price index).

Details:
This change is pretty straightforward - replace GetPaginationEntityByUtilization with GetEntityStats. Note that in the GetEntityStats call for a global entity type we don't need to provide the candidate IDs, and the GetEntityStats call doesn't currently accept an entity type because we figure out the entity types on the backend (this can - and maybe should - change in the future, to get rid of an unnecessary DB query in the history component).

Testing Done:
- Full build
- System test - started XL and performed searches with sorting by utilization. Verified entities returned by history component come in order of price index when requested.

Reviewed By: zhengbin
https://rbcommons.com/s/VMTurbo/r/24875/