OM-36437: XL Repository: In plans, saving price index information alongside the projected topologies in ArangoDB.
Intent:
To be able to sort plan entity stats by price index.

Background:
We don't currently save price indices on an entity basis anywhere for plans. The history component only stores the aggregate information. We need this information for some UI widgets to work - e.g. the optimized improvement widget, which relies on a sort by price index in the projected topologies.

Changes:
I'm simply storing the ProjectedTopologyEntity instead of the TopologyEntityDTO in the collections created by TopologyProtobufWriter, and retrieving them in TopologyProtobufReader. This is the most straightforward approach, and it's good enough for our purposes. Pagination is still done in memory, but the cost of moving pagination to the database is too high given the current use cases (we need to change the way we store the topology information).

To deal with backwards compatibility, the TopologyProtobufReader first attempts to read ProjectedTopologyEntity objects from the topology collection. If a single read fails, it switches to try to read TopologyEntityDTO objects. We assume that all entities in a topology collection will either be TopologyEntityDTO objects (saved before this change) or ProjectedTopologyEntity objects (saved after this change), and not a mix of both, so we can save the overhead of parsing attempts and exception handling - which can be significant when dealing with lots of entities!

Testing Done:
- Full build
- System testing - started XL and ran a plan, verified optimized widgets render.

Reviewed By: psoohoo, mark.laff
https://rbcommons.com/s/VMTurbo/r/24820/