OM-36653 - Store probeId along with entityDescriptor in TP assigned_identity table. RB:24869
Add a new column to the TP assigned_identity table to store the ID of the probe
which discovered the entity. This is useful when we want to migrate the
EntityDescriptors when the IdentityMetadata changes. Without this information,
it is not possible to filter the entities that need to be upgraded when the
IdentityMetadata changes.