OM-37509: Fix broken XL performance tests RB:24998
* See http://jnm01.eng.vmturbo.com:8080/job/xl-trunk-performance-tests/1222/

The MEDIATION_UCSDIRECTOR_XMX_MB variable is not set. Defaulting to a blank string.
The MEDIATION_UCSDIRECTOR_SYSTEM_PROPERTIES variable is not set. Defaulting to a blank string.
The MEDIATION_UCSDIRECTOR_MEM_LIMIT_MB variable is not set. Defaulting to a blank string.
The MEDIATION_UCSDIRECTOR_PORT variable is not set. Defaulting to a blank string.
The MEDIATION_UCSDIRECTOR_DEBUG_PORT variable is not set. Defaulting to a blank string.
The Compose file '/automation/workspace/XL/build/docker-compose.yml' is invalid because...

Testing Done:
* Successfully ran ActionOrchestratorPerformanceTests locally after fixing.