OM-12900 Graceful shutdown of components. RB: 25226
Fix the SIGTERM handling in the components during docker stop. "docker stop"
sends SIGTERM signal to the container. If there is no response in 10
seconds(default timeout), it sends SIGKILL to the container. Most of the times,
we do docker stop. So if we fix this issue, then the components(including
mysqld) would be shutdown gracefully and this would reduce the probability of
corruption(due to arbitrarily killing the components).

* Propogate signals to child process in docker containers
* Add shutdown hook to java components