OM-37129: XL TP: Setting capacity increment from probe-provided increment, and adjusting user resize increment to units of probe-provided increment.
Issue:
We need to account for probe-provided resize increments for commodities, and reconcile mismatched user and probe-provided resize increments.

Background:
The SDK supports the ability of probes to send resize increments for commodities. Right now we don't take those increments into account, and the resize increments in TopologyEntityDTO are derived exclusively from user settings. This is a problem, because users may set resize increments that are not acceptable by the probe.

Solution:
1) Set the capacity increment in commodities from the increment sent by the probe.
2) When processing the resize setting, round it down to the nearest multiple of the probe-provided increment.

Note: The probes currently do NOT send the resize increments. That should be handled separately.

Testing Done:
- Full build.
- Manually injected probe increment of 1024, and verified that user increment of 1 or 1038 gets rounded to 1024, and 2050 gets rounded to 2048. The probe doesn't currently send increments.

Reviewed By: mark.laff
https://rbcommons.com/s/VMTurbo/r/25231/