OM-37324: XL: Define pricing tables (cost matrix) that will be initialized from the pricing data discovered by the cost probe.
This change defines the pricing table that will be used as one of the inputs into the cost calculation library in both the cost component and the market component. The other inputs will be VM <-> RI mappings and discounts (as well as the topology to calculate cost for).

The pricing table format is different than what's discovered by the cost probe. The cost probe discovers "identifying information" - for example, the names of the tiers, the names of the regions they're in, and so on. The topology processor is then responsible for:
- Converting the names (and any other identifying information) to IDs - the same IDs that we'll derive from the objects discovered by the billing probe and the topology probe.
- Formatting the data into the PricingTable format.
- Uploading the PricingTable to the cost component.
    - Note: The topology processor may also need to upload the entity information to the cost component (e.g. the name of the template), so that if the market recommends changing a VM to use a different template discovered by the cost component we have a way of getting display information for the template ID.

The work in the topology processor, and the cost component RPC's to upload the information, will be done in a separate review. The work to convert the AWS and Azure cost probes to the SDK format will also be done separately.

Additional changes:
- Moving enums/messages shared between RIs and Pricing into Common.proto.
- Adding support for currency into the messages.

Testing Done:
- Build.

Reviewed By: dblinn, emanuele.maccherani
https://rbcommons.com/s/VMTurbo/r/25113/