OM-37783 - Cancel queued automated actions when the automation settings are changed by the user. RB:25305
Automation actions(such as move, resize etc) when set to execute automatically,
are queued and executed one by one. If the pace of action execution is slow(by
the probe), the actions may be in the queue for a long time. Meanwhile, if the
Automation setting policies are changed by the user via the UI, it should be
reflected immediately. So we cancel all the tasks in the queue which have not
yet been executed wheneve the automation settings changes. For simplicity, we
purge/cancel all actions instead of purging/cancelling only those actions whose
entitiy setting policies were changed.

Added a new gRPC call in AO called cancelQueuedActions() which will be
triggered whenever any of the AUTOMATION settings are changed.