OM-37725: Create library for OID Store RB-25282: Roman, Christine, David, Wen
Background
Topology Processor has logic to implement a process for comparing the attributes of two Entities to determine if they are the same, and to manage the OID creation and persistence for Entities.

Problem
We would like to implement a similar algorithm to create/persist OIDs for other types of objects we discover.

Solution
We've created a new library to implement the desired framework:  map from a generic item to Attributes; then look up or create an OID using the attributes as a key.