OM-38255: conform XL to new field name in ReportScheduleApiDTO RB-25558: Roman
background
The ReportScheduleApiDTO structure is used to return info about a report schedule.

problem
the field for the unique ie was originally misspelled - subcriptionId instead of subscriptionId

solution
rename the field internally, and add a new @Deprecated field with the old (incorrect) name that returns the value of the correctly-named field.

Needs to be reflected in XL