OM-38096: XL Cost: Add interface to put/get price tables into the cost component.
Intent:
The cost component needs to be able to receive price tables discovered by cost probes, and needs to make that pricing data externally accessible for use by the market component (for the Market's cost calculation).

Background:
The cloud workflow requires that the pricing data (i.e. the price tables for each cost probe) be stored in the cost component. The cost calculation library harnesses in the market and the cost component will then retrieve these pricing tables as one of the inputs into the cost calculation algorithm.

Adding a GRPC service to upload the probe-type-specific cost tables, and to retrieve the merged price table for cost calculation. The TP will use the service when processing discovery results of the cost probe. The Market component will use the service when getting the price table for it's cost calculation. The cost component won't use the service, but it will use the underlying PriceTableStore.

Note: The current implementation keeps the price tables in memory. There is a separate story to keep them in the database, but that part is less important for now until we have the whole workflow completed.

Testing Done:
- Full build.
- Started up Cost component, added and retrieved price table via Swagger.

Reviewed By: psoohoo
https://rbcommons.com/s/VMTurbo/r/25521