OM-37222: Persist Workflow items in Action Orchestrator RB-25493: Roman
Background
Workflows are non-market-entities discovered from Orchestration-type targets such as UCS-Director. These Workflows will be stored on the Action Orchestrator, and then when set as the "external action" for an Automation Policy, Action Orchestrator will direct the action to the Orchestration target instead of the target that discovered the original SE.

Approach
In this change we add two new tables ('workflow_oid' and 'workflow') to the Action Orchestrator partition in MariaDB ('action') to hold the Workflow oids and the related workflow items are written to the 'workflow' table.. The 'workflow_oid' table has a primary part key - an OID - which is generated based on two of the workflow object charactaristics: the orchestration target id from which the workflow was discovered, and the external name of the workflow object. The OID generation requires a persistent store to record the prior (attributes -> OID) mappings we have assigned. We use the 'workflow_oid' table for that. When a 'workflow' object is uploaded, we look up the (target-id, workflow-name) in an in-memory cache; if found, we use the prior OID; if not found, we generate a new OID and store the mapping from (target-id, workflow-name) -> new-oid in the persistent store. Further, we use the "last-modified" timestamp in the 'workflow' table to detect when a previously uploaded Workflow is no longer part of the upload from a given target. In that case we delete that row from 'workflow_oid' and, using

We also use the 'workflow' table to store the original WorkflowInfo protobuf as a blob. We will use this in future stories to provide the original workflow fields to the UI and to the action execution steps.

recommit