OM-38173 - Purge queued actions which are not recommended again by the market in the current broadcast cycle. RB:25650
Due to slow execution of actions, the actions pile up in the execution queue.
Actions can be executed several hours after the maket had recommended that
action. This leads to execution of actions which are no longer relevant and it
shows up as errors in the UI. To fix this issue, we retain IN_PROGRESS, QUEUED
and READY actions which are re-recommended by the market. This way only current
and up-to-date actions are executed by AO.

Also added a new action state transition from QUEUED -> CLEARED when QUEUED
actions are cancelled.