OM-37220:  Enable adding "Orchestration" Settings Policy (Automation Policies) RB-25681: Christine, Emanuele, Roman, PengCheng
background
In XL the settings emulated by the Group component, defined by the EntitySettingsSpec enum, instead of Manager classes as in Classic.

This implementation should enable the three settings for which Orchestration currently defined, in Classic and in the UCS-D Probe:
- provisionPMWorkflow - provision a pm
- suspendPMWorkflow - suspend a pm
- deleteSTWorkflow - delete a storage

In XL, there is no longer a 'delete' action, so that setting is not needed in XL.

In API component there's a mapping to simulate the Managers on which the /settingspolicies is built, based on the settingManagers.json file. Each entry defines a manager in the Classic sense, including the display name, category to which the settings belong, and a list of the settings the manager manages. Each setting also defines a list of the entity types to which the setting applies, so there is no longer a need for identifying the entity type in the setting name.

solution
We add three two enum values to EntitySettingSpec, one for each of the desired orchestration settings. We call them: provisionActionWorkflow and suspendActionWorkflow, where the name implies "this is the workflow to invoke on behalf of an ___ action". The "value" of each of these settings is simply the external_name of the orchestration script to be run (on the Orchestration target).

We also add an entry to settingManagers.json for the 'controlmanager', listing these three settings as the ones managed, so that the UI will be able to retrieve the settings/policies and auto-generate the UI for them.

We also moved some common code, createCategoryPath(), into a 'util' class, used by both EntitySettingsSpec and GlobalSettingsSpec.