OM-37224: Implement the /workflows endpoint in the API Component RB-25790: Christine, Jeff, Patrick, Roman
**background**
In order to implement the "Action Orchestration", Add Action Policy function, the UI requests the discovered Workflow objects using the /workflows endpoint. See the attached screen shots. These workflows are stored in the Action Orchestrator component, and so the API component must fetch the workflows from there for return to the UI.

**solution**
- Implement the WorkflowsService to the API component to handle the IWorkflowsService method for 'getWorkflows()'; the 'getWorkflowByByUuid()' is not required for this task
- Create an IWorkflowsService bean in api component ServiceConfig.java
- Add a FetchWorkflowsRequest and FetchWorkflowsService to WorkflowDTO.proto
- Add a FetchWorkflowRpcService to ActionOrchestrator
- Add a 'fetchWorkflows()' method to PersistentWorkflowStore

Note that there's a bug in the UI (OM 38669) which request an actionscriptmanager entry if __any__ controlmanager (Orchestration) policies are to be viewed. This required temporary entries in settingManagers.json in the API component and EntitySettingSpecs.java in components.common