OM-37327: XL: Cost component integration with cost calculation library.
Also addresses 38097 and 38675.

Intent: To apply cost calculation library to cloud entities in the topology coming in to the cost component.

The main change is to read the entities from the remote iterator and save them to a map. The "TopologyCostCalculator" class just initializes the integration classes, and calls the CloudCostCalculator for every TopologyEntityDTO in the cloud topology.

Extending the EntityCostStore to accept a list of CostJournal objects and write into the entity_cost table from those objects directly. This is better than converting the journals to protobufs just to send them to EntityCostStore. Also, chunking the entity cost storage operation.

Testing Done:
- Full build; unit tests pass.
- Additional SQLEntityCostStore tests.
- Started cost component.

Reviewed By: mark.laff, garyzeng
https://rbcommons.com/s/VMTurbo/r/25791/