OM-38843: Test failure in Plan Orchestrator ActionListenerTest RB-25891: Patrick, David
background
The tests in com.vmturbo.plan.orchestrator.plan depend on a config file PlanTestConfig.java. This config includes an @PostConstruct which does initialization common to all the tests the include this config class.

problem
There is a test failure in Jenkins for 7.5.0 which we've never seen before. The message says "Table 'plan_test.plan_instance' doesn't exist" during flyway migration.

We cannot be sure this is the root cause, but there was no corresponding @PreDesctroy, so these tests never cleaned up after themselves. In particular, the "flyway.clean()" was never executed to remo ve the temp database.

solution
Added an @PreDestroy to execute the "flyway.clean()", and @DirtiesContext to each test to ensure that this cleanup method is called.

Also, PlanDaoImplTest was needlessly importing TestSQLDatabaseConfig which is already imported by PlanTestConfig.java.