OM-38817: XL: When retrieving plan per-entity stats, return no stats when scope is empty.
Right now an empty scope will actually return all entities, which isn't what we want in the per-entity-stat case.

In the long term, we may want to return all entities in the API, but handle the empty-scope case in the UI (removing the query if there are no entities to get results for). However, the fix in this review is the fastest for 7.4.0.

Testing Done:
- Ran a plan scoped to a datacenter. The "Storage" optimized improvement widget no longer has VMs in the "after" bars.

Reviewed By: mark.laff, karthikt
https://rbcommons.com/s/VMTurbo/r/25928/