OM-38770: Implement /workflows/{workflow-id} endpoint RB-25849: Roman, Christine, Jeff, Ioannis
background
The UI calls the endpoint /workflows/{workflow-id} to display information about an Orchestration policy that the user created previously. This endpoint was not yet implemented in XL at all.

fix
We add
(1) an implementation of the getWorkflowByUuid() method in WorkflowsService in the API Component that calls a new service added to Workflows.proto
(2) add a new gRPC endpoint in the WorkflowService to "FetchWorkflow"
(3) implement the new gRPC endpoint in WorkflowService
(4) implement a new method in PersistentWorkflowStore

We also modified PersistentWorkflowStore to use a Jooq Conversion method to store/retrieve a WorkflowInfo object to/from the database (where it is saved in a blob column).