OM-38833: Implement workflow indirection when returning actions to the UI RB-25955: Jeff
Background
Action execution is subject to Policy control. In the case of Orchestration (Workflow) policies, the user may define a policy for an action, such as "provision of a PM", and an ActionMode (DISABLED, RECOMMEND, MANUAL, AUTOMATIC) which controls both how the Action is displayed in the UI as well as when (if ever) the action should be executed. In particular, if an action type is in RECOMMEND mode the action is displayed but without a selection "checkbox", so the user has no way to execute the action. In DISABLED mode the action isn't displayed at all, and in AUTOMATIC mode execution begins "right away".

task
If there is a workflow policy defined (we only support three: Provision Host, Suspend Host, and Suspend Storage) then we ignore the action capabilities from the original probe (e.g. VC), which may not allow Suspend, for example, and use the ActionMode from the settings that the user defines, both global and per group.

details
We add a method 'calculateWorkflowActionMode' to check for the 'workflow action' setting and if so calculate the ActionMode as above.  If no 'workflow' setting then the ActionMode is calculated as before, i.e. gated by the probe's action capabilities.