OM-38855:  Ensure 'displayName' field is set in TargetApiDTO in XL RB-25993: Yannis, Christine
background
The 'displayName' property returned in a 'TargetApiDTO' is supposed to be set from the 'address' account value. The UX typically displays the 'address' account value if the 'displayName' is not set, with the exception of the "Workflow" selection list. Without the 'displayName', the Workflow selection list shows 'undefined' (see first screen shot).
Note that the target account value 'nameOrAddress' is also accepted. This was the "old" account value used for the same purpose as 'address'.

fix
Added code to the mapper for Target -> TargetApiDTO in the API Component which looks for the target account value 'address' or 'nameOrAddress' and, if found, set the 'displayName' in the return TargetApiDTO.

... sorry for the two-part commit ...