OM-38173 : Fix RecommendationTracker Iterator. RB:26058
RecommendationTracker structure is used to track re-recommended actions. Any old
action which remains in the tracker after the actions have been "taken" by
matching with the new actions, has to be cleared. We were iterating the
RecommendationTracker and clearing them one by one. But due to a bug in the
iterator, we missing clearing of many actions.