OM-38834: Implement workflow indirection during action execution RB-26065: Jeff
background
If the user defines a Setting policy for an Orchestration action, that identifies a Workflow that should be executed on an Orchestration Target (e.g. UCS-Director) instead of the action being overridden.  For example, the user might define an Orchestration Policy for "provisionWorkflow" on a PhysicalMachine. In that case, any "provision" action on that PhysicalMachine will instead be as a "provisionWorkflow", and so directed to UCS-D for execution instead of the (e.g.) VC from which the PM was discovered.

Further, the Workflow to be invoked is passed to the Orchestration Probe/Target in the 'workflow' field of the SDK ActionExecutionDTO.

solution
In the Action Orchestrator, we check for a Setting defining a Workflow for the given Action. If found, we build the ActionExecution.proto request using the target id from the original Action target id (based on where the entity was discovered) to the target for the corresponding Workflow, and include the correponding Workflow object. This ActionExecution proto is sent to the Topology Processor.

In Topology Processor OperationManager.requestActions() we create a new NonMarketEntityDTO correponding to the WorkflowInfo and add that to the ActionRequest to be passed to the probe. For this to work we needed to relax the "all entities discovered by the same target" rule - the "target" for this operation is the Orchestration provider, not the original hypervisor. Jeff provided the intended patch (unfortunately this broke a test for exactly that rule - so the test was @Ignored for now).

Also, we needed to add a case to ActionExecutionRpcService.executAction() to handle a PROVISION action. One did not previously exist.