OM-38946: XL Metrics: Making MetricWarehouseTest more resistant to concurrency issues.
Intent:
The tests should pass even in a slow environment (relative to the collection interval).

Background:
In a slow environment - or if you set the collection interval to 10ms in a fast environment - the warehouse would sometimes scrape more than the expected number of times ( another scrape would happen between the first scrape and the call to shutdownNow()).

There is no fix required to the actual warehouse - because an extra scrape is not a problem in actual use. For the tests to succeed, we de-dupe the collected "TimestampedSample"s before comparing them against the expected samples. Since we always control the clock and the value being collected, the de-duped samples should be equal.

Testing Done:
- Ran the unit test with a 1ms collection interval. It failed before, and does not fail anymore.

Reviewed By: dblinn
https://rbcommons.com/s/VMTurbo/r/26031/