OM-38945: XL AO: Translating actions when counting, so that action counts align with action retrieval results.
Intent
======
Translate actions before counting them for the various counting RPC calls. This way we're less likely to count actions that fail translation.

Background
===================
We often get mismatched action counts and numbers of action actions returned from the GetActions calls. This was because actions that fail/get discarded in translation - most notably VCPU resizes where the to and from numbers are the same - appear in the counts but not in the actual actions returned.

The solution is to apply action translation when counting actions. This means the action counts calls will be slower than before - but not by much, since successful translations are cached, and translations are already executed in batch.

Note: I'm moving functionality out of the QueryFilter for two reasons:
1. The ActionTranslator needs to work with a stream of ActionViews, and all the additional grouping manipulations should be done after translation.
2. The query filter should just be responsible for the filtering. The grouping of filtered results should be done outside the filter.

Testing Done:
- New unit tests (+ refactoring due to moving functionality out of QueryFilter).

Reviewed By: garyzeng
https://rbcommons.com/s/VMTurbo/r/26062/