OM-39143: Implement CPU Template processing in the API Component RB-26095: Joseph, Roman
background
We have added a field to Templates for Physical Machines (Hosts) for a 'cpuModel' string. The 'cpuModel' is looked up from a list of CPU's presented by the com.turbonomic.cpucapacity.CPUCatalog. This 'cpuModel' should be stored when the Template is created or edited, and returned when the template is queried.

There is an external REST API endpoint to query the available CPU models.

implementation
We added the com.turbonomic.cpucapacity library to the API component dependencies and implemented the endpoint Service and Mapper for the CPU model structure (CpuModelApiDTO).
We also added a field 'cpu_model' to the TemplateInfo message in TemplateDTO.proto.