OM-36961: XL Search: Allow case insensitive string property filter.
Intent
======
Search (over names in particular) should be case insensitive in XL.

Background
==========
We currently match case when resolving string property filters during searches in both Repository and Topology Processor. In classic pretty much all string-based searches are case-insensitive.

I'm extending the StringFilter with an option to match case - or not. Right now all creations of StringFilter in the API component will use the "case-insensitive" option. In the future we may want to do a case-sensitive search if the user input contains a capital letter.

In the repository, I'm switching from the equality/regex operators to the REGEX_TEST() function to do case-sensitive or case-insensitive search. Also, simplifying some of the method calls to create the AQL string directly from the StringFilter/MapFilter.

Testing Done:
- Full build, with additional unit tests in TopologyFilterFactoryTest.
- Make sure search is case-insensitive via the UI. (this tests that the Repository is performing the search correctly)
    - Also mixed-case search (to make sure search text with uppercase letters matches lowercase results)
- Created a dynamic group based on the search, created a policy to set actions to "RECOMMEND". Verified that after the next broadcast entities in the group have non-clickable actions. (this tests that the TP is performing the search correctly)

Reviewed By: zhengbin
https://rbcommons.com/s/VMTurbo/r/26089/