OM-38783 - Upgrade vCenter VM entity identity records stored in TP with new heuristic property. RB: 25979
The heuristic properties for vCenter VM entities was changed by this commit in
the SDK: https://rbcommons.com/s/VMTurbo/r/24982/diff/5/#2. In XL, we have to upgrade
the probeInfo and the entity identifying properties to include these changes.

The upgrade is done in two phases:
1) During the 1st phase, the probe info is updated to include the new fields and
the entities are updated to have the new heuristic property with a dummy value.
As we currently don't store entityType, we can't decide which entity to update.
So we add a new entity_type column to the the assigned_identity table and then
populate it by fetching the entityTypes from history component. For those entities
in the assigned_identity table for which we couldn't find the entityType in
history component, we delete them.
2) In the 2nd phase, during the normal discovery of the entities, when the
heuristic properties match, we replace the dummy value with the actual value.

The migration added a new flyway script and it's version number(V_2_3) is the same
as the version number for the create_targetspec_table script. Since the
create_targetspec_table script was added in develop(higher version) branch,
we are renaming this flyway script to have a higher version number.