OM-39312: XL Group: Preserving clusterHeadroomTemplateId when updating discovered Cluster.
Issue
=====
The TargetCollectionUpdate for clusters did not preserve the cluster headroom template ID, so any modifications to it would be lost on the next topology broadcast.

Solution
========
Set the cluster headroom template ID of the updated cluster object to the value in the existing cluster object in the Group component.

Testing
=======
New unit test - failed before, passes now.

Reviewed By: pankajnarula, psoohoo
https://rbcommons.com/s/VMTurbo/r/26216/