OM-39351: Add CPU Capacity library to Plan Orchestrator RB-26292: Patrick, Yannis
background
We have created a library CPU Capacity which estimates a "performance scale factor" based a CPU model name and the SPEC benchmark results. This "scale factor" is applied prior to market analysis to better calculate CPU Capacity - i.e. this is more accurate than simply multiplying MHz * #cores.

implementation
This functionality is closely related to Templates, which are hosted in the Plan Orchestrator. For that reason, we are locating the CPU Performance Estimation library there. There are two main functions of this library: return a list of all CPU Model names we know of, so that the user may pick from a drop-down in the UI;  and for a given list of CPU model names, return a map from each CPU model name to the corresponding "estimated scale factor".

This step simply adds the gRPC service for each of the two desired functions. The following reviews will make use of these functions in the API Component and Topology Processor component.